+++ pokepi/src/models/encounter_condition_values.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct EncounterConditionValue {
-    pub encounter_condition_value_id: u16,
+    pub encounter_condition_value_id: i32,
     pub value_name: String,
     pub value_identifier: String,
 }
