+++ pokepi/src/models/pokedex_entry.rs
@@ -2,10 +2,10 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct PokedexEntry {
-    pub national_id: u16,
-    pub form_id: Option<u16>,
-    pub game_id: u16,
-    pub pokedex_number: u16,
+    pub national_id: i32,
+    pub form_id: Option<i32>,
+    pub game_id: i32,
+    pub pokedex_number: i32,
     pub entry_text: String,
 }
 
@@ -13,6 +13,6 @@ pub type CreatePokedexEntry = PokedexEntry;
 
 #[derive(Debug, Deserialize, Clone)]
 pub struct UpdatePokedexEntry {
-    pub pokedex_number: Option<u16>,
+    pub pokedex_number: Option<i32>,
     pub entry_text: Option<String>,
 }
