+++ pokepi/src/models/pokemon_form_ability.rs
@@ -2,9 +2,9 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct PokemonFormAbility {
-    pub pokemon_form_game_id: u32,
-    pub ability_id: u16,
-    pub slot: u8,
+    pub pokemon_form_game_id: i32,
+    pub ability_id: i32,
+    pub slot: i16,
     pub is_hidden: bool,
 }
 
@@ -12,6 +12,6 @@ pub type CreatePokemonFormAbility = PokemonFormAbility;
 
 #[derive(Debug, Deserialize, Clone)]
 pub struct UpdatePokemonFormAbility {
-    pub slot: Option<u8>,
+    pub slot: Option<i16>,
     pub is_hidden: Option<bool>,
 }
