+++ pokepi/src/models/version_group.rs
@@ -2,22 +2,22 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct VersionGroup {
-    pub version_group_id: u16,
+    pub version_group_id: i32,
     pub version_group_name: String,
     pub version_group_identifier: String,
-    pub generation: u8,
+    pub generation: i16,
 }
 
 #[derive(Debug, Deserialize, Clone)]
 pub struct CreateVersionGroup {
     pub version_group_name: String,
     pub version_group_identifier: String,
-    pub generation: u8,
+    pub generation: i16,
 }
 
 #[derive(Debug, Deserialize, Clone)]
 pub struct UpdateVersionGroup {
     pub version_group_name: Option<String>,
     pub version_group_identifier: Option<String>,
-    pub generation: Option<u8>,
+    pub generation: Option<i16>,
 }
